api/requirements.txt
pymongo==4.6.1
flask==3.0.0
jsonschema==4.21.1
python-dotenv==1.0.1
mysql-connector-python==8.3.0
mariadb==1.1.10